+++ RELEASE_NOTES.md
@@ -32,12 +32,15 @@ The technology is effectively dead and the project no longer actively maintained
 In prevent compilation errors inside the workspace, those components have been
 removed from the project.
 
-### Changes to wbp-component.xml schema
+### Changes to wbp-component.xml & wbp-factory.xml schema
 
 All special characters in the description need to be properly encoded.
 
 Example:
 &lt;b&gt;...&lt;/b&gt;
+
+The wbp-factory.xml files should use the http://www.eclipse.org/wb/WBPComponent
+namespace.
 
 ## 1.14.0 (2023-12)